+++ wemake_python_styleguide/visitors/ast/complexity/overuses.py
@@ -2,22 +2,20 @@
 
 import ast
 from collections import defaultdict
-from typing import ClassVar, DefaultDict, List, Union
+from typing import Callable, ClassVar, DefaultDict, List, Tuple
 
 from typing_extensions import final
 
 from wemake_python_styleguide.compat.aliases import FunctionNodes
-from wemake_python_styleguide.constants import SPECIAL_ARGUMENT_NAMES_WHITELIST
-from wemake_python_styleguide.logic import nodes, source, walk
+from wemake_python_styleguide.logic import source, walk
+from wemake_python_styleguide.logic.complexity import overuses
 from wemake_python_styleguide.types import AnyNodes
 from wemake_python_styleguide.violations import complexity
 from wemake_python_styleguide.visitors import base
 
+#: We use these types to store the number of nodes usage in different contexts.
 _Expressions = DefaultDict[str, List[ast.AST]]
 _FunctionExpressions = DefaultDict[ast.AST, _Expressions]
-_Annotated = Union[ast.arg, ast.AnnAssign]
-
-_AnnNodes = (ast.AnnAssign, ast.arg)
 
 
 @final
@@ -41,27 +39,7 @@ class StringOveruseVisitor(base.BaseNodeVisitor):
         self.generic_visit(node)
 
     def _check_string_constant(self, node: ast.Str) -> None:
-        is_type_annotaion = False
-
-        annotated = walk.get_closest_parent(node, (*_AnnNodes, *FunctionNodes))
-        if isinstance(annotated, FunctionNodes):
-            is_type_annotaion = bool(
-                node == annotated.returns or
-                (
-                    annotated.returns and
-                    walk.is_contained_by(node, annotated.returns)
-                ),
-            )
-        elif isinstance(annotated, _AnnNodes):
-            is_type_annotaion = bool(
-                node == annotated.annotation or
-                (
-                    annotated.annotation and
-                    walk.is_contained_by(node, annotated.annotation),
-                ),
-            )
-
-        if is_type_annotaion:
+        if overuses.is_annotation(node):
             return
 
         self._string_constants[node.s] += 1
@@ -89,7 +67,6 @@ class ExpressionOveruseVisitor(base.BaseNodeVisitor):
         ast.BinOp,
         ast.Call,
         ast.Compare,
-        ast.Starred,
         ast.Subscript,
         ast.UnaryOp,
         ast.Lambda,
@@ -102,6 +79,15 @@ class ExpressionOveruseVisitor(base.BaseNodeVisitor):
         ast.GeneratorExp,
         ast.Set,
         ast.SetComp,
+    )
+
+    _ignore_predicates: Tuple[Callable[[ast.AST], bool], ...] = (
+        overuses.is_decorator,
+        overuses.is_self,
+        overuses.is_annotation,
+        overuses.is_class_context,
+        overuses.is_super_call,
+        overuses.is_primitive,
     )
 
     _msg: ClassVar[str] = '{0}; used {1}'
@@ -127,18 +113,7 @@ class ExpressionOveruseVisitor(base.BaseNodeVisitor):
         self.generic_visit(node)
 
     def _add_expression(self, node: ast.AST) -> None:
-        ignore_predicates = [
-            self._is_decorator,
-            self._is_self_method,
-            self._is_annotation,
-
-            # We use this predicate because classes have quite complex
-            # DSL to be created: like django-orm, attrs, and dataclasses.
-            # And these DSLs are built using attributes and calls.
-            _is_class_context,
-            _is_super_call,
-        ]
-        if any(ignore(node) for ignore in ignore_predicates):
+        if any(ignore(node) for ignore in self._ignore_predicates):
             return
 
         source_code = source.node_to_string(node)
@@ -150,39 +125,6 @@ class ExpressionOveruseVisitor(base.BaseNodeVisitor):
                 node,
             )
 
-    def _is_decorator(
-        self,
-        node: ast.AST,
-    ) -> bool:
-        parent = walk.get_closest_parent(node, FunctionNodes)
-        if isinstance(parent, FunctionNodes) and parent.decorator_list:
-            return any(
-                node == decorator or walk.is_contained_by(node, decorator)
-                for decorator in parent.decorator_list
-            )
-        return False
-
-    def _is_self_method(self, node: ast.AST) -> bool:
-        if isinstance(node, ast.Call) and isinstance(node.func, ast.Attribute):
-            if isinstance(node.func.value, ast.Name):
-                if node.func.value.id in SPECIAL_ARGUMENT_NAMES_WHITELIST:
-                    return True
-        return False
-
-    def _is_annotation(self, node: ast.AST) -> bool:
-        typed_assign = walk.get_closest_parent(
-            node,
-            (ast.AnnAssign, ast.arg),
-        )
-
-        if isinstance(typed_assign, _AnnNodes) and typed_assign.annotation:
-            is_same_node = node == typed_assign.annotation
-            is_child_annotation = walk.is_contained_by(
-                node, typed_assign.annotation,
-            )
-            return is_same_node or is_child_annotation
-        return False
-
     def _post_visit(self) -> None:
         for mod_source, module_nodes in self._module_expressions.items():
             if len(module_nodes) > self.options.max_module_expressions:
@@ -190,6 +132,7 @@ class ExpressionOveruseVisitor(base.BaseNodeVisitor):
                     complexity.OverusedExpressionViolation(
                         module_nodes[0],
                         text=self._msg.format(mod_source, len(module_nodes)),
+                        baseline=self.options.max_module_expressions,
                     ),
                 )
 
@@ -203,13 +146,3 @@ class ExpressionOveruseVisitor(base.BaseNodeVisitor):
                             baseline=self.options.max_function_expressions,
                         ),
                     )
-
-
-def _is_class_context(node: ast.AST) -> bool:
-    return isinstance(nodes.get_context(node), ast.ClassDef)
-
-
-def _is_super_call(node: ast.AST) -> bool:
-    if isinstance(node, ast.Call) and isinstance(node.func, ast.Name):
-        return node.func.id == 'super'
-    return False